// proto/user-service/users/user.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.7.5
//   protoc               v5.29.3
// source: proto/user-service/users/user.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
import { Observable } from "rxjs";
import { Timestamp } from "../../../google/protobuf/timestamp";

export const protobufPackage = "user.v1";

export interface User {
  id: string;
  firstName: string;
  lastName?: string | undefined;
  email: string;
  primaryPhoneNumber: string;
  phoneNumbers: string[];
  password: string;
  createdAt: Timestamp | undefined;
  updatedAt: Timestamp | undefined;
  deletedAt: Timestamp | undefined;
}

export interface CreateUserRequest {
  firstName: string;
  lastName: string;
  email: string;
  primaryPhoneNumber: string;
  password: string;
}

export interface CreateUserResponse {
  status: string;
  message: string;
  data: User | undefined;
  errors: string[];
}

export interface FindAllUsersRequest {
  page: string;
  limit: string;
}

export interface FindAllUsersResponse {
  users: User[];
}

export interface FindOneUserRequest {
  id: string;
}

export interface FindOneUserResponse {
  user: User | undefined;
}

export interface UpdateUserRequest {
  id: string;
  firstName: string;
  lastName: string;
  email: string;
  primaryPhoneNumber: string;
}

export interface UpdateUserResponse {
  user: User | undefined;
}

export interface RemoveUserRequest {
  id: string;
}

export interface RemoveUserResponse {
  message: string;
}

export interface QueryUsersRequest {
  page: number;
  limit: number;
}

export interface QueryUsersResponse {
  users: User[];
}

export const USER_V1_PACKAGE_NAME = "user.v1";

export interface UserServiceClient {
  createUser(request: CreateUserRequest): Observable<CreateUserResponse>;

  findAllUsers(request: FindAllUsersRequest): Observable<FindAllUsersResponse>;

  findOneUser(request: FindOneUserRequest): Observable<FindOneUserResponse>;

  updateUser(request: UpdateUserRequest): Observable<UpdateUserResponse>;

  removeUser(request: RemoveUserRequest): Observable<RemoveUserResponse>;

  queryUsers(request: Observable<QueryUsersRequest>): Observable<QueryUsersResponse>;
}

export interface UserServiceController {
  createUser(
    request: CreateUserRequest,
  ): Promise<CreateUserResponse> | Observable<CreateUserResponse> | CreateUserResponse;

  findAllUsers(
    request: FindAllUsersRequest,
  ): Promise<FindAllUsersResponse> | Observable<FindAllUsersResponse> | FindAllUsersResponse;

  findOneUser(
    request: FindOneUserRequest,
  ): Promise<FindOneUserResponse> | Observable<FindOneUserResponse> | FindOneUserResponse;

  updateUser(
    request: UpdateUserRequest,
  ): Promise<UpdateUserResponse> | Observable<UpdateUserResponse> | UpdateUserResponse;

  removeUser(
    request: RemoveUserRequest,
  ): Promise<RemoveUserResponse> | Observable<RemoveUserResponse> | RemoveUserResponse;

  queryUsers(request: Observable<QueryUsersRequest>): Observable<QueryUsersResponse>;
}

export function UserServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ["createUser", "findAllUsers", "findOneUser", "updateUser", "removeUser"];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod("UserService", method)(constructor.prototype[method], method, descriptor);
    }
    const grpcStreamMethods: string[] = ["queryUsers"];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod("UserService", method)(constructor.prototype[method], method, descriptor);
    }
  };
}

export const USER_SERVICE_NAME = "UserService";